testing separate files for functional components
