+++ front-end/CopyPasta/App.js
@@ -1,12 +1,12 @@
 import { StatusBar } from "expo-status-bar";
 import React from "react";
 import { StyleSheet, Text, View } from "react-native";
-import { ConnectedDevices } from "./app";
+import ConnectedDevices from "./app/ConnectedDevices";
 
 export default function App() {
   return (
     <View style={styles.container}>
-      <Text>Hello</Text>
+      <ConnectedDevices />
     </View>
   );
 }
